Addressing a bug with links in Game Details not working
Removed the links to the high and low spread plays as they would be inaccuate if there were multiple highs and lows.  Changed the behavior of the other hig/low links to show a list of plays that match that rather than trying to link to a single play which wouldn't always be the only one.